Remove AddExitProc and replace with try ... finally block.
AROS does not allow for an ExitProc procedure to be invoked when normal
program termination happens as SDL libraries are already unloaded at that
point (exit procedures are called _after_ PASCALMAIN ended).

This patch from ALB42 solves abnormal program termination with a normal
flow for AROS target and at the same time is compatible for the other
supported targets.